DELETE /movies/{movieId}/character/{characterId} working!

+++ disney-api/src/main/java/com/alkemy/disneyapi/service/impl/MovieServiceImpl.java
@@ -1,11 +1,15 @@
 package com.alkemy.disneyapi.service.impl;
 
+import com.alkemy.disneyapi.dto.CharacterDTO;
 import com.alkemy.disneyapi.dto.MovieBasicDTO;
 import com.alkemy.disneyapi.dto.MovieDTO;
+import com.alkemy.disneyapi.entity.CharacterEntity;
 import com.alkemy.disneyapi.entity.MovieEntity;
 import com.alkemy.disneyapi.exception.ParamNotFound;
+import com.alkemy.disneyapi.mapper.CharacterMapper;
 import com.alkemy.disneyapi.mapper.MovieMapper;
 import com.alkemy.disneyapi.repository.MovieRepository;
+import com.alkemy.disneyapi.service.CharacterService;
 import com.alkemy.disneyapi.service.MovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,12 +24,17 @@ public class MovieServiceImpl implements MovieService {
     private MovieRepository movieRepository;
     @Autowired
     private MovieMapper movieMapper;
+    @Autowired
+    private CharacterService characterService;
+    @Autowired
+    private CharacterMapper characterMapper;
 
     //GET
     public List<MovieBasicDTO> getAllMoviesLite() {
         List<MovieEntity> entitiesList = movieRepository.findAll();
         return movieMapper.movieEntityList2BasicDTOList(entitiesList);
     }
+
     public MovieDTO getMovieDetailsById(Long id) {
         Optional<MovieEntity> movieEntity = movieRepository.findById(id);
         if (!movieEntity.isPresent()) {
@@ -54,9 +63,30 @@ public class MovieServiceImpl implements MovieService {
 
     }
 
-    @Override
+    //DELETE
     public void deleteMovie(Long id) {
-        movieRepository.deleteById(id);
+        if (movieRepository.existsById(id)) {
+            movieRepository.deleteById(id);
+        } else {
+            throw new ParamNotFound("Movie ID not exists in Database");
+        }
+    }
+
+    public void deleteCharacterFromMovie(Long movieId, Long characterId) {
+        MovieEntity movie = getMovieEntityById(movieId);
+        movie.getCharacters().size();
+        CharacterEntity characterEntity = characterService.getCharacterEntityById(characterId);
+        movie.removeCharacterFromMovie(characterEntity);
+        movieRepository.save(movie);
+
+    }
+
+    public MovieEntity getMovieEntityById(Long movieId) {
+        Optional<MovieEntity> movie = movieRepository.findById(movieId);
+        if (!movie.isPresent()) {
+            throw new ParamNotFound("Movie Id not found");
+        }
+        return movie.get();
     }
